requirements.txt
MetaTrader5
pandas
loguru
google-genai
python-dotenv
ta